git-svn-id: file:///svn/toku/tokudb@18181 c7de825b-a66e-492c-adef-691d508d4ae1

+++ toku_include/toku_portability.h
@@ -135,7 +135,8 @@ void  os_free(void*);
 // full_pwrite and full_write performs a pwrite, and checks errors.  It doesn't return unless all the data was written. */
 void toku_os_full_pwrite (int fd, const void *buf, size_t len, toku_off_t off) __attribute__((__visibility__("default")));
 void toku_os_full_write (int fd, const void *buf, size_t len) __attribute__((__visibility__("default")));
-ssize_t toku_os_write (int fd, const void *buf, size_t len) __attribute__((__visibility__("default")));
+// os_write returns 0 on success, otherwise an errno.
+int toku_os_write (int fd, const void *buf, size_t len) __attribute__((__visibility__("default")));
 
 // wrapper around fsync
 int toku_file_fsync(int fd);
